+++ heat/sparse/__init__.py
@@ -1,7 +1,7 @@
 """add sparse heat function to the ht.sparse namespace"""
 
 from .arithmetics import *
-from .dcsr_matrix import *
+from .dcsx_matrix import *
 from .factories import *
 from ._operations import *
 from .manipulations import *
