+++ heat/sparse/dcsx_matrix.py
@@ -343,6 +343,26 @@ class DCSR_matrix(__DCSX_matrix):
     """
     Distributed Compressed Sparse Row Matrix. It is composed of
     PyTorch sparse_csr_tensors local to each process.
+
+    Parameters
+    ----------
+    array : torch.Tensor (layout ==> torch.sparse_csr)
+        Local sparse array
+    gnnz: int
+        Total number of non-zero elements across all processes
+    gshape : Tuple[int,...]
+        The global shape of the array
+    dtype : datatype
+        The datatype of the array
+    split : int or None
+        If split is not None, it denotes the axis on which the array is divided between processes.
+        DCSR_matrix only supports distribution along axis 0.
+    device : Device
+        The device on which the local arrays are using (cpu or gpu)
+    comm : Communication
+        The communications object for sending and receiving data
+    balanced: bool or None
+        Describes whether the data are evenly distributed across processes.
     """
 
     @property
@@ -364,6 +384,26 @@ class DCSC_matrix(__DCSX_matrix):
     """
     Distributed Compressed Sparse Column Matrix. It is composed of
     PyTorch sparse_csc_tensors local to each process.
+
+    Parameters
+    ----------
+    array : torch.Tensor (layout ==> torch.sparse_csc)
+        Local sparse array
+    gnnz: int
+        Total number of non-zero elements across all processes
+    gshape : Tuple[int,...]
+        The global shape of the array
+    dtype : datatype
+        The datatype of the array
+    split : int or None
+        If split is not None, it denotes the axis on which the array is divided between processes.
+        DCSR_matrix only supports distribution along axis 0.
+    device : Device
+        The device on which the local arrays are using (cpu or gpu)
+    comm : Communication
+        The communications object for sending and receiving data
+    balanced: bool or None
+        Describes whether the data are evenly distributed across processes.
     """
 
     @property
